Move config/computed values out. Convert single capture screen as well.
