+++ lib/models/constants.dart
@@ -1,0 +1,3 @@
+const flashStartDuration = Duration(milliseconds: 50);
+const flashEndDuration = Duration(milliseconds: 2500);
+const minimumContinueWait = Duration(milliseconds: 1500);
