+++ lib/views/photo_booth_screen/screens/capture_screen/capture_screen_view_model.dart
@@ -1,23 +1,15 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:mobx/mobx.dart';
-import 'package:momento_booth/hardware_control/gphoto2_camera.dart';
-import 'package:momento_booth/hardware_control/photo_capturing/live_view_stream_snapshot_capturer.dart';
 import 'package:momento_booth/hardware_control/photo_capturing/photo_capture_method.dart';
-import 'package:momento_booth/hardware_control/photo_capturing/sony_remote_photo_capture.dart';
 import 'package:momento_booth/main.dart';
-import 'package:momento_booth/managers/live_view_manager.dart';
-import 'package:momento_booth/managers/mqtt_manager.dart';
 import 'package:momento_booth/managers/photos_manager.dart';
-import 'package:momento_booth/managers/project_manager.dart';
 import 'package:momento_booth/managers/settings_manager.dart';
 import 'package:momento_booth/managers/stats_manager.dart';
-import 'package:momento_booth/models/capture_state.dart';
+import 'package:momento_booth/models/constants.dart';
 import 'package:momento_booth/models/maker_note_data.dart';
-import 'package:momento_booth/models/settings.dart';
 import 'package:momento_booth/views/base/screen_view_model_base.dart';
 import 'package:momento_booth/views/components/imaging/photo_collage.dart';
 import 'package:momento_booth/views/photo_booth_screen/screens/share_screen/share_screen.dart';
@@ -31,21 +23,14 @@ abstract class CaptureScreenViewModelBase extends ScreenViewModelBase with Store
   late final PhotoCaptureMethod capturer;
   bool flashComplete = false;
   bool captureComplete = false;
-  static const flashStartDuration = Duration(milliseconds: 50);
-  static const flashEndDuration = Duration(milliseconds: 2500);
-  static const minimumContinueWait = Duration(milliseconds: 1500);
 
   int get counterStart => getIt<SettingsManager>().settings.captureDelaySeconds;
-  int get autoFocusMsBeforeCapture => getIt<SettingsManager>().settings.hardware.gPhoto2AutoFocusMsBeforeCapture;
 
   double get collageAspectRatio => getIt<SettingsManager>().settings.collageAspectRatio;
   double get collagePadding => getIt<SettingsManager>().settings.collagePadding;
 
   @computed
-  Duration get photoDelay => Duration(seconds: counterStart) - capturer.captureDelay + flashStartDuration;
-
-  @computed
-  Duration get autoFocusDelay => photoDelay - Duration(milliseconds: autoFocusMsBeforeCapture);
+  Duration get photoDelay => getIt<PhotosManager>().photoDelay;
 
   @observable
   bool showCounter = true;
@@ -96,18 +81,7 @@ abstract class CaptureScreenViewModelBase extends ScreenViewModelBase with Store
   CaptureScreenViewModelBase({
     required super.contextAccessor,
   }) {
-    capturer = switch (getIt<SettingsManager>().settings.hardware.captureMethod) {
-      CaptureMethod.sonyImagingEdgeDesktop => SonyRemotePhotoCapture(getIt<SettingsManager>().settings.hardware.captureLocation),
-      CaptureMethod.liveViewSource => LiveViewStreamSnapshotCapturer(),
-      CaptureMethod.gPhoto2 => getIt<LiveViewManager>().gPhoto2Camera!,
-    };
-    capturer.clearPreviousEvents();
-
-    if (autoFocusMsBeforeCapture > 0 && autoFocusDelay > Duration.zero && capturer is GPhoto2Camera) {
-      Future.delayed(autoFocusDelay).then((_) => (capturer as GPhoto2Camera).autoFocus());
-    }
-    Future.delayed(photoDelay).then((_) => captureAndGetPhoto());
-    getIt<MqttManager>().publishCaptureState(CaptureState.countdown);
+    getIt<PhotosManager>().initiateDelayedPhotoCapture(onCaptureFinished);
   }
 
   Future<void> onCounterFinished() async {
@@ -121,28 +95,9 @@ abstract class CaptureScreenViewModelBase extends ScreenViewModelBase with Store
     navigateAfterCapture();
   }
 
-  Future<void> captureAndGetPhoto() async {
-    getIt<MqttManager>().publishCaptureState(CaptureState.capturing);
-
-    try {
-      final image = await capturer.captureAndGetPhoto();
-      getIt<StatsManager>().addCapturedPhoto();
-      getIt<PhotosManager>().photos.add(image);
-      if (getIt<ProjectManager>().settings.singlePhotoIsCollage) {
-        await captureCollage();
-      } else {
-        getIt<PhotosManager>().outputImage = image.data;
-        await getIt<PhotosManager>().writeOutput();
-      }
-    } catch (error) {
-      logWarning(error);
-      final ByteData data = await rootBundle.load('assets/bitmap/capture-error.png');
-      getIt<PhotosManager>().outputImage = data.buffer.asUint8List();
-    } finally {
-      captureComplete = true;
-      navigateAfterCapture();
-      getIt<MqttManager>().publishCaptureState(CaptureState.idle);
-    }
+  void onCaptureFinished() {
+    captureComplete = true;
+    navigateAfterCapture();
   }
 
   void navigateAfterCapture() {
